requirements.txt
fastparquet==2024.11.0
gensim==4.3.3
graphviz==0.20.3
ipdb==0.13.13
matplotlib==3.10.1
networkx==3.3
numba==0.61.0
numpy==2.2.3
ogb==1.3.6
pandas==2.2.3
Pillow==11.1.0
pip==24.0
pretrainedmodels==0.7.4
pyarrow==19.0.1
PyYAML==6.0.1
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==69.5.1
skimage==0.0
tensorboardX==2.6.2.2
tensorly==0.9.0
texttable==1.7.0
torch==2.3.1
torch_geometric==2.6.1
torch_scatter==2.1.2
torch_sparse==0.6.18
torchvision==0.21.0
torchviz==0.0.3
tqdm==4.66.2
wheel==0.45.1
